+++ src/index.ts
@@ -26,3 +26,6 @@ export * from './withParams'
 
 // CORS
 export * from './cors'
+
+// TYPES
+export * from './types'
